Made some changes to the cleanDBTables function
Before, the cleanDBTables function went through and deleted the tables in the order of whats in the DataBaseObjectType enum. The problem with this is that it'll delete the values in the location table before the equipment table. Equipment has location as a foreign key so that threw some sql errors. I just made a array of DBObjectTypes with location below everything else. The delete sql line was just "DELETE" and the toString value of the enum which is not the name of the table. For example, ComputerSR.toString is "Computer Service Request" and the table name is ComputerSR. I assumed that toString value is used somewhere else so I made a new function that'll return the table name. Also, the sql line should have been "DELETE FROM" so i just changed that.